new system of rating

+++ dev/clubmein/scripts/directives/itemDirective.js
@@ -35,18 +35,27 @@ var item = ['$state','RootService','MembersService','$rootScope','$mdDialog','$t
 					scope.bContent = true;
 			};
 
-			scope.loadBids = function(){
-				if(!scope.bids){
-					var request = clubService.sendRequest('GET', '/item/'+scope.item.itemId+'/bids');
-					request.then(function(response){
-						console.log('bids ---- ', response);
-						scope.bids = response;
-						scope.bBidsSection = true;
+			scope.showBids = function(){
+				if(scope.bBidsContent)
+					scope.bBidsContent = false;
+				else
+					scope.bBidsContent = true;
+					
+			};
+
+			scope.bidDetails = function(bid){
+				if(bid){
+					//create record that the user clicked on see bid details
+					var request = clubService.sendRequest('PUT', '/item/'+scope.item.itemId+'/record', bid.bidId);
+					request.then( function(response){
+						console.log('response::: ', response);
+						$state.go('app.bidInformation', {id: bid.bidId});
 					}, function(error){
-						clubService.sendRequest('error getting bids', 'error');
+						clubService.addNotification('error, couldnt get bid information', 'error');
 					});
+					
 				}
-			}
+			};
 
 			scope.addOther = function(parent, type){
 				var array = [];
@@ -247,6 +256,8 @@ var item = ['$state','RootService','MembersService','$rootScope','$mdDialog','$t
 						city: scope.city
 					};
 					scope.item.location = scope.location;
+				}else{
+					scope.item.location = null;
 				}
 				if(bSave){
 					if(scope.isItemRepeated(scope.item)){
@@ -260,8 +271,7 @@ var item = ['$state','RootService','MembersService','$rootScope','$mdDialog','$t
 							
 							scope.item.itemId= response.itemId;
 							scope.item.location = response.location;
-							scope.safeItem.itemId = response.itemId;
-							scope.safeItem.location = response.location;
+							scope.safeItem = angular.copy(scope.item);
 							clubService.addNotification('added item successfully', 'success');
 							scope.bContent = false;
 						}, function(error){
@@ -289,9 +299,14 @@ var item = ['$state','RootService','MembersService','$rootScope','$mdDialog','$t
 
 			scope.cancel = function (){
 				scope.bContent = false;
-				scope.item = angular.copy(scope.safeItem);
+			 	if(scope.item.itemId){
+						scope.item = angular.copy(scope.safeItem);
+			 	}else{
+			 		scope.item = null;
+			 	}
 
 				console.log('item cancel', scope.item);
+				scope.bBidsContent = false;
 			};
 
 			scope.showContent = function(){
@@ -309,8 +324,9 @@ var item = ['$state','RootService','MembersService','$rootScope','$mdDialog','$t
 
 			scope.otherLocation = function(){
 				scope.item.otherLocation != scope.item.otherlocation;
-				//console.log('scope.item.otherLocation', scope.item.otherLocation);
+				
 			};
+
 
 			scope.showConfirm = function(ev, code, type){
 				 // Appending dialog to document.body to cover sidenav in docs app
@@ -331,14 +347,18 @@ var item = ['$state','RootService','MembersService','$rootScope','$mdDialog','$t
 		    
 			};
 			scope.deleteItem = function(){
-				var request = clubService.sendRequest('DELETE', '/item/'+scope.item.itemId);
-				request.then(function(response){
-					membersService.removeItem(response);
+				if(scope.item.itemId){
+					var request = clubService.sendRequest('DELETE', '/item/'+scope.item.itemId);
+					request.then(function(response){
+						membersService.removeItem(response);
+						scope.item = null;
+						scope.bContent = false;
+					}, function(error){
+						clubService.addNotification('error deleting the item', 'error');
+					});
+				}else{
 					scope.item = null;
-					scope.bContent = false;
-				}, function(error){
-					clubService.addNotification('error deleting the item', 'error');
-				});
+				}
 			};
